# src/config/swagger.json
{
  "openapi": "3.0.3",
...
  "info": {
    "title": "Auth-User-API",
    "description": "Authentication user API",
    "contact": {
      "name": "Paulo Dantas",
      "url": "https://github.com/paulodantas-dev",
      "email": "paulodantasbjr@gmail.com"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:5000/",
      "description": "Development server"
    },
    {
      "url": "https://auth-app-node.herokuapp.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/refresh_token": {
      "post": {
        "description": "Returns user and token",
        "summary": "Authenticate user",
        "operationId": "refresh_token",
        "responses": {
          "200": {
            "description": "refresh_token response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "400": {
            "description": "Please login now."
          },
          "401": {
            "description": "Token invalid."
          },
          "404": {
            "description": "This does not exist."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Response": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "fullname": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "email": {
                "type": "string",
                "format": "email"
              }
            }
          }
        },
        "required": ["token", "user"]
      },
...
    }
  }
}